Factor out Ansible utils and host discovery
Extract two separate modules from rho/scancommand.py. This is part of
the preparation for scanning for Fuse.
