+++ rho/scancommand.py
@@ -7,7 +7,6 @@
 # FOR A PARTICULAR PURPOSE. You should have received a copy of GPLv2
 # along with this software; if not, see
 # http://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.
-
 # pylint: disable=too-few-public-methods, too-many-lines
 
 """ Rho CLI Commands """
@@ -15,53 +14,21 @@
 from __future__ import print_function
 import os
 import sys
-import re
 import time
 import json
 from tempfile import NamedTemporaryFile
-from collections import defaultdict
-from getpass import getpass
-import pexpect
-import yaml
 from rho import utilities
 from rho.clicommand import CliCommand
 from rho.vault import get_vault_and_password
 from rho.utilities import (
-    multi_arg, _read_in_file, str_to_ascii, iteritems, PING_LOG_PATH,
-    ANSIBLE_LOG_PATH, SCAN_LOG_PATH, PING_INVENTORY_PATH, PROFILE_HOSTS_SUFIX,
-    PROFILE_HOST_AUTH_MAPPING_SUFFIX, log
+    multi_arg, _read_in_file, str_to_ascii, iteritems,
+    SCAN_LOG_PATH, PROFILE_HOSTS_SUFIX,
+    PROFILE_HOST_AUTH_MAPPING_SUFFIX
 )
+from rho import host_discovery
 from rho import facts
+from rho import ansible_utils
 from rho.translation import _
-
-
-def auth_as_ansible_host_vars(auth):
-    """Get the Ansible host vars that implement an auth.
-
-    :param auth: the auth. A dictionary with fields 'id', 'name',
-        'username', 'password', and 'ssh_key_file'.
-
-    :returns: a dict that can be used as the host variables in an
-        Ansible inventory.
-    """
-
-    username = auth.get('username')
-    password = auth.get('password')
-    ssh_key_file = auth.get('ssh_key_file')
-    sudo_password = auth.get('sudo_password')
-
-    ansible_vars = {}
-
-    ansible_vars['ansible_user'] = str_to_ascii(username)
-    if password:
-        ansible_vars['ansible_ssh_pass'] = str_to_ascii(password)
-    if ssh_key_file:
-        ansible_vars['ansible_ssh_private_key_file'] = \
-            str_to_ascii(ssh_key_file)
-    if sudo_password:
-        ansible_vars['ansible_become_pass'] = sudo_password
-
-    return ansible_vars
 
 
 def redacted_auth_string(auth):
@@ -80,167 +47,6 @@ def redacted_auth_string(auth):
 
     return '{0}, {1}, {2}, {3}'.format(
         name, username, utilities.PASSWORD_MASKING, ssh_key_file)
-
-
-def redact_dict(redact_key_list, a_dict):
-    """Redact_values in a dictionary
-
-    :param redact_key_list: keys in dictionary that value should be redacted
-    :param a_dict: A dictionary to redact
-    :returns: the redacted dictionary
-    """
-    for key in redact_key_list:
-        if a_dict is not None and key in a_dict:
-            a_dict[key] = utilities.PASSWORD_MASKING
-    return a_dict
-
-
-def log_yaml_inventory(label, inventory):
-    """Log yaml inventory but mask passwords
-
-    :param inventory: A dictionary of the ansible inventory
-    """
-    alpha = inventory.get('alpha')
-    hosts_dict = alpha.get('hosts')
-    vars_dict = alpha.get('vars')
-    redact_key_list = ['ansible_become_pass', 'ansible_ssh_pass']
-
-    # pylint: disable=unused-variable
-    for host, host_dict in iteritems(hosts_dict):
-        host_dict = redact_dict(redact_key_list, host_dict)
-
-    vars_dict = redact_dict(redact_key_list, vars_dict)
-
-    log.debug('%s:\n%s', label, yaml.dump(inventory))
-    return inventory
-
-
-def process_ping_output(out_lines):
-    """Find successful hosts from the output of a ping command.
-
-    Use this function by using ansible to run echo "Hello" on remote
-    hosts, then sending the output to this function.
-
-    :param out_lines: an iterator returning lines of Ansible output.
-    :returns: the hosts that pinged successfully, as a set and those that
-        failed, as a set.
-    """
-
-    success_hosts = set()
-    failed_hosts = set()
-
-    # Ansible output has the format
-    # host | UNREACHABLE! => {
-    #     "changed": false,
-    #     "msg": "Failed to connect to the host via ssh ...",
-    #     "unreachable": true
-    #    }
-    #   hostname | SUCCESS | rc=0 >>
-    #   Hello
-    # with the above two lines repeated for each host
-    for line in out_lines:
-        ansi_escape = re.compile(r'\x1b[^m]*m')
-        line = ansi_escape.sub('', line)
-        pieces = line.split('|')
-        if len(pieces) == 3 and pieces[1].strip() == 'SUCCESS':
-            success_hosts.add(pieces[0].strip())
-        elif len(pieces) == 3 and pieces[1].strip() == 'FAILED':
-            failed_hosts.add(pieces[0].strip())
-        elif len(pieces) == 2 and pieces[1].strip().startswith('UNREACHABLE'):
-            failed_hosts.add(pieces[0].strip())
-
-    log.debug('Ping log reached hosts: %s', success_hosts)
-    log.debug('Ping log did not reached hosts: %s', failed_hosts)
-
-    return success_hosts, failed_hosts
-
-
-# Creates the inventory for pinging all hosts and records
-# successful auths and the hosts they worked on
-# pylint: disable=too-many-statements, too-many-arguments, unused-argument
-def _create_ping_inventory(vault, vault_pass, profile_ranges, profile_port,
-                           credential, forks, ansible_verbosity):
-
-    """Find which auths work with which hosts.
-
-    :param vault: a Vault object
-    :param vault_pass: password for the Vault?
-    :param profile_ranges: hosts for the profile
-    :param profile_port: the SSH port to use
-    :param credential: auth to use
-    :param forks: the number of Ansible forks to use
-
-    :returns: a tuple of
-      (list of IP addresses that worked for any auth,
-       map from host IPs to SSH ports that worked with them,
-       map from host IPs to lists of auths that worked with them
-      )
-    """
-
-    # pylint: disable=too-many-locals
-    success_hosts = set()
-    failed_hosts = set()
-    success_port_map = defaultdict()
-    success_auth_map = defaultdict(list)
-    hosts_dict = {}
-
-    for profile_range in profile_ranges:
-        # pylint: disable=anomalous-backslash-in-string
-        reg = "[0-9]*.[0-9]*.[0-9]*.\[[0-9]*:[0-9]*\]"
-        profile_range = profile_range.strip(',').strip()
-        hostname = str_to_ascii(profile_range)
-        if not re.match(reg, profile_range):
-            hosts_dict[profile_range] = {'ansible_host': profile_range,
-                                         'ansible_port': profile_port}
-        else:
-            hosts_dict[hostname] = None
-
-    vars_dict = auth_as_ansible_host_vars(credential)
-
-    yml_dict = {'alpha': {'hosts': hosts_dict, 'vars': vars_dict}}
-    vault.dump_as_yaml_to_file(yml_dict, PING_INVENTORY_PATH)
-    log_yaml_inventory('Ping inventory', yml_dict)
-
-    print(_('Attempting connection discovery with auth "%s".' %
-            (credential.get('name'))))
-
-    cmd_string = 'ansible alpha -m raw' \
-                 ' -i ' + PING_INVENTORY_PATH \
-                 + ' --ask-vault-pass -f ' + forks \
-                 + ' -a \'echo "Hello"\''
-
-    my_env = os.environ.copy()
-    my_env["ANSIBLE_HOST_KEY_CHECKING"] = "False"
-    # Don't pass ansible_verbosity here as adding too much
-    # verbosity can break our parsing of Ansible's output. This is
-    # a temporary fix - a better solution would be less-fragile
-    # output parsing.
-    run_ansible_with_vault(cmd_string, vault_pass,
-                           log_path=PING_LOG_PATH,
-                           env=my_env,
-                           log_to_stdout=True,
-                           ansible_verbosity=0)
-
-    with open(PING_LOG_PATH, 'r') as ping_log:
-        success_hosts, failed_hosts = process_ping_output(ping_log)
-
-    for host in success_hosts:
-        success_auth_map[host].append(credential)
-        success_port_map[host] = profile_port
-
-    num_success = len(success_hosts)
-    num_failed = len(failed_hosts)
-    if num_success > 0:
-        print(_('Connection succeeded with auth "%s" to %d systems.') %
-              (credential.get('name'), num_success))
-    if num_failed > 0:
-        print(_('Failed to connect with auth "%s" to %d systems.') %
-              (credential.get('name'), num_failed))
-    if num_success > 0 or num_failed > 0:
-        print()
-
-    return list(success_hosts), success_port_map, success_auth_map, \
-        list(failed_hosts)
 
 
 # Helper function to create a file to store the mapping
@@ -308,7 +114,8 @@ def make_inventory_dict(success_hosts, success_port_map, auth_map):
         ascii_port = str_to_ascii(str(success_port_map[host]))
         host_vars = {'ansible_host': ascii_host,
                      'ansible_port': ascii_port}
-        host_vars.update(auth_as_ansible_host_vars(auth_map[host][0]))
+        host_vars.update(
+            ansible_utils.auth_as_ansible_host_vars(auth_map[host][0]))
         alpha_hosts[ascii_host] = host_vars
 
     yml_dict['alpha'] = {'hosts': alpha_hosts}
@@ -322,88 +129,7 @@ def _create_main_inventory(vault, success_hosts, success_port_map,
     hosts_yml = profile + PROFILE_HOSTS_SUFIX
     hosts_yml_path = utilities.get_config_path(hosts_yml)
     vault.dump_as_yaml_to_file(yml_dict, hosts_yml_path)
-    log_yaml_inventory('Main inventory', yml_dict)
-
-
-def run_ansible_with_vault(cmd_string, vault_pass, env=None, log_path=None,
-                           log_to_stdout=True, ansible_verbosity=0):
-    """ Runs ansible command allowing for password to be provided after
-    process triggered.
-
-    Returns after the process completes.
-
-    :param cmd_string: the command to run.
-    :param vault_pass: the password to the user's Ansible Vault.
-    :param env: the environment to run the subprocess in.
-    :param log_path: a path to write the process's log to. Defaults to
-        'XDG_DATA_HOME/rho/ansible_log'.
-    :param log_to_stdout: if True, write Ansible's log to stdout. Defaults to
-        True.
-    :param ansible_verbosity: the number of v's of Ansible verbosity.
-    :returns: the popen.spawn object for the process.
-    """
-
-    # pexpect provides the ability to send the process's output to a
-    # single Python file object. We want to send it to a file and
-    # maybe also stdout. The solution is to have pexpect log to the
-    # file and then use 'tail -f' to copy that to stdout.
-
-    if not log_path:
-        log_path = ANSIBLE_LOG_PATH
-
-    if ansible_verbosity:
-        cmd_string = cmd_string + ' -' + 'v' * ansible_verbosity
-
-    result = None
-    try:
-        utilities.ensure_data_dir_exists()
-        with open(log_path, 'wb') as logfile:
-            pass
-        with open(log_path, 'r+b') as logfile:
-            log.debug('Running Ansible: %s', cmd_string)
-            child = pexpect.spawn(cmd_string, timeout=None,
-                                  env=env)
-
-            if log_to_stdout:
-                utilities.threaded_tailing(log_path, ansible_verbosity)
-
-            result = child.expect('Vault password:')
-            child.sendline(vault_pass)
-            # Set the log file *after* we send the user's Vault
-            # password to Ansible, so we don't log the password.
-            child.logfile = logfile
-            last_pos = logfile.tell()
-
-            i = child.expect([pexpect.EOF, 'Enter passphrase for key .*:',
-                              'you want to continue connecting (yes/no)?'])
-            while i:
-                new_pos = logfile.tell()
-                logfile.seek(last_pos)
-                logfile_lines = logfile.readlines()
-                log.info(logfile_lines)
-                print(logfile_lines[-1].replace('\r\n', ''))
-                logfile.seek(new_pos)
-                last_pos = new_pos
-                child.logfile = None
-                # Ansible has already printed a prompt; it would be
-                # confusing if getpass printed another one.
-                child.sendline(getpass(''))
-                child.logfile = logfile
-                i = child.expect([pexpect.EOF, 'Enter passphrase for key .*:',
-                                  'you want to continue connecting (yes/no)?'])
-
-            if child.isalive():
-                child.wait()
-            if log_to_stdout:
-                time.sleep(2)
-
-        return child
-    except pexpect.EOF:
-        print(str(result))
-        print('pexpect unexpected EOF')
-    except pexpect.TIMEOUT:
-        print(str(result))
-        print('pexpect timed out')
+    ansible_utils.log_yaml_inventory('Main inventory', yml_dict)
 
 
 class ScanCommand(CliCommand):
@@ -501,6 +227,7 @@ class ScanCommand(CliCommand):
             self.parser.print_help()
             sys.exit(1)
 
+    # pylint: disable=too-many-statements
     def _do_command(self):
         # pylint: disable=too-many-locals
         # pylint: disable=too-many-branches
@@ -562,11 +289,12 @@ class ScanCommand(CliCommand):
             for cred_item in profile_auth_list:
                 success_hosts_, success_port_map_, \
                     auth_map_, remaining_hosts_ = \
-                    _create_ping_inventory(vault, vault_pass,
-                                           remaining_hosts,
-                                           profile_port,
-                                           cred_item, forks,
-                                           self.verbosity)
+                    host_discovery.create_ping_inventory(
+                        vault, vault_pass,
+                        remaining_hosts,
+                        profile_port,
+                        cred_item, forks,
+                        self.verbosity)
                 success_hosts = success_hosts + success_hosts_
                 remaining_hosts = remaining_hosts_
                 success_port_map.update(success_port_map_)
@@ -635,10 +363,11 @@ class ScanCommand(CliCommand):
         else:
             log_path = SCAN_LOG_PATH
         print('Running:', cmd_string)
-        process = run_ansible_with_vault(cmd_string, vault_pass,
-                                         log_path=log_path,
-                                         log_to_stdout=True,
-                                         ansible_verbosity=self.verbosity)
+        process = ansible_utils.run_with_vault(
+            cmd_string, vault_pass,
+            log_path=log_path,
+            log_to_stdout=True,
+            ansible_verbosity=self.verbosity)
         if process.exitstatus == 0 and process.signalstatus is None:
             host_auth_mapping = \
                 self.options.profile + PROFILE_HOST_AUTH_MAPPING_SUFFIX
